Crate alloy_consensus

alloy-consensus

Ethereum consensus interface.

This crate contains constants, types, and functions for implementing Ethereum EL consensus and communication. This includes headers, blocks, transactions, EIP-2718 envelopes, EIP-2930, EIP-4844, and more.

In general a type belongs in this crate if it is committed to in the EL block header. This includes:

transactions

blocks

headers

receipts

EIP-2718 envelopes.

Provenance

Much of this code was ported from reth-primitives as part of ongoing alloy migrations.

Modules

constants

Ethereum protocol-related constants

eip4844_utils

Utilities for working with EIP-4844 field elements and implementing SidecarCoder.

header 🔒

receipt 🔒

sealed 🔒

signed 🔒

transaction 🔒

Structs

Blob

A basic blob data.

BlobTransactionSidecar

This represents a set of blobs, and its corresponding commitments and proofs.

Bytes48

An array of 48 bytes. Represents an untrusted (potentially invalid) commitment/proof.

Header

Ethereum Block header

Receipt

Receipt containing result of transaction execution.

ReceiptWithBloom

Receipt with calculated bloom filter.

Sealed

A consensus hashable item, with its memoized hash.

SidecarBuilder

Build a BlobTransactionSidecar from an arbitrary amount of data.

Signed

A transaction with a signature and hash seal.

SimpleCoder

Simple coder that only uses the last 31 bytes of each blob. This is the default coder for the SidecarBuilder.

TxEip1559

A transaction with a priority fee (EIP-1559).

TxEip2930

Transaction with an [AccessList] (EIP-2930).

TxEip4844

EIP-4844 Blob Transaction

TxEip4844WithSidecar

EIP-4844 Blob Transaction

TxLegacy

Legacy transaction.

Enums

BlobTransactionValidationErrorkzg

An error that can occur when validating a TxEip4844Variant.

EnvKzgSettingskzg

KZG settings.

ReceiptEnvelope

Receipt envelope, as defined in EIP-2718.

TxEip4844Variant

EIP-4844 Blob Transaction

TxEnvelope

The Ethereum EIP-2718 Transaction Envelope.

TxType

Ethereum TransactionType flags as specified in EIPs 2718, 1559, and 2930.

TypedTransaction

The TypedTransaction enum represents all Ethereum transaction request types.

Constants

EMPTY_OMMER_ROOT_HASH

Ommer root of empty list.

EMPTY_ROOT_HASH

Root hash of an empty trie.

Traits

Sealable

Sealeable objects.

SidecarCoder

A strategy for coding and decoding data into sidecars. Coder instances are responsible for encoding and decoding data into and from the sidecar. They are called by the SidecarBuilder during the ingest, take, and (if c_kzg feature enabled) build methods.

SignableTransaction

A signable transaction.

Transaction

Represents a minimal EVM transaction.

TxReceipt

Receipt is the result of a transaction execution.