Struct alloy_consensus::receipt::receipts::ReceiptWithBloom

pub struct ReceiptWithBloom<T = Log> {
    pub receipt: Receipt<T>,
    pub logs_bloom: Bloom,
}

Receipt with calculated bloom filter.

This convenience type allows us to lazily calculate the bloom filter for a receipt, similar to Sealed.

Fields
receipt: Receipt<T>

The receipt.

logs_bloom: Bloom

The bloom filter.

Implementations

impl<T: Encodable> ReceiptWithBloom<T>

fn receipt_rlp_header(&self) -> Header

Returns the rlp header for the receipt payload.

fn encode_fields(&self, out: &mut dyn BufMut)

Encodes the receipt data.

fn payload_len(&self) -> usize

impl<T> ReceiptWithBloom<T>

pub const fn new(receipt: Receipt<T>, logs_bloom: Bloom) -> Self

Create new ReceiptWithBloom

pub fn into_components(self) -> (Receipt<T>, Bloom)

Consume the structure, returning the receipt and the bloom filter

fn decode_receipt(buf: &mut &[u8]) -> Result<Self>
where T: Decodable,

Decodes the receipt payload