Enum alloy_consensus::request::Request

#[non_exhaustive]
pub enum Request {
    DepositRequest(DepositRequest),
    WithdrawalRequest(WithdrawalRequest),
    ConsolidationRequest(ConsolidationRequest),
}

Ethereum execution layer requests.

See also EIP-7685.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

DepositRequest(DepositRequest)

An EIP-6110 deposit request.

WithdrawalRequest(WithdrawalRequest)

An EIP-7002 withdrawal request.

ConsolidationRequest(ConsolidationRequest)

An EIP-7251 consolidation request.

Implementations

impl Request

pub const fn is_deposit_request(&self) -> bool

Whether this is a [DepositRequest].

pub const fn is_withdrawal_request(&self) -> bool

Whether this is a [WithdrawalRequest].

pub const fn is_consolidation_request(&self) -> bool

Whether this is a [ConsolidationRequest].

pub const fn as_deposit_request(&self) -> Option<&DepositRequest>

Return the inner [DepositRequest], or None of this is not a deposit request.

pub const fn as_withdrawal_request(&self) -> Option<&WithdrawalRequest>

Return the inner [WithdrawalRequest], or None if this is not a withdrawal request.

pub const fn as_consolidation_request(&self) -> Option<&ConsolidationRequest>

Return the inner [ConsolidationRequest], or None if this is not a consolidation request.