alloy_consensus

Struct Sealed

pub struct Sealed<T> { /* private fields */ }

A consensus hashable item, with its memoized hash.

We do not implement any specific hashing algorithm here. Instead types implement the Sealable trait to provide define their own hash.

Implementations

impl<T> Sealed<T>

pub fn new(inner: T) -> Sealed<T>
where T: Sealable,

Seal the inner item.

pub fn new_ref(inner: &T) -> Sealed<&T>
where T: Sealable,

Seal the inner item, by reference.

pub fn new_with<F>(inner: T, f: F) -> Sealed<T>
where F: FnOnce(&T) -> FixedBytes<32>,

Seal the inner item with some function.

pub fn new_ref_with<F>(inner: &T, f: F) -> Sealed<&T>
where F: FnOnce(&T) -> FixedBytes<32>,

Seal a reference to the inner item with some function.

pub const fn new_unchecked(inner: T, seal: FixedBytes<32>) -> Sealed<T>

Instantiate without performing the hash. This should be used carefully.

pub const fn as_sealed_ref(&self) -> Sealed<&T>

Converts from &Sealed<T> to Sealed<&T>.

pub fn into_parts(self) -> (T, FixedBytes<32>)

Decompose into parts.

pub fn split(self) -> (T, FixedBytes<32>)

Decompose into parts. Alias for Self::into_parts.

pub fn clone_inner(&self) -> T
where T: Clone,

Clone the inner item.

pub const fn inner(&self) -> &T

Get the inner item.

pub const fn seal(&self) -> FixedBytes<32>

Get the hash.

pub const fn hash(&self) -> FixedBytes<32>

Get the hash.

pub const fn hash_ref(&self) -> &FixedBytes<32>

Get the hash.

pub fn into_inner(self) -> T

Unseal the inner item, discarding the hash.

pub fn unseal(self) -> T

Unseal the inner item, discarding the hash. Alias for Self::into_inner.

impl<T> Sealed<&T>

pub fn cloned(self) -> Sealed<T>
where T: Clone,

Maps a Sealed<&T> to a Sealed<T> by cloning the inner value.