alloy_consensus
Trait BlockHeader

pub trait BlockHeader {

    // Required methods
    fn parent_hash(&self) -> B256;
    fn ommers_hash(&self) -> B256;
    fn beneficiary(&self) -> Address;
    fn state_root(&self) -> B256;
    fn transactions_root(&self) -> B256;
    fn receipts_root(&self) -> B256;
    fn withdrawals_root(&self) -> Option<B256>;
    fn logs_bloom(&self) -> Bloom;
    fn difficulty(&self) -> U256;
    fn number(&self) -> BlockNumber;
    fn gas_limit(&self) -> u64;
    fn gas_used(&self) -> u64;
    fn timestamp(&self) -> u64;
    fn mix_hash(&self) -> B256;
    fn nonce(&self) -> B64;
    fn base_fee_per_gas(&self) -> Option<u64>;
    fn blob_gas_used(&self) -> Option<u64>;
    fn excess_blob_gas(&self) -> Option<u64>;
    fn parent_beacon_block_root(&self) -> Option<B256>;
    fn requests_root(&self) -> Option<B256>;
    fn extra_data(&self) -> &Bytes;

}

Trait for extracting specific Ethereum block data from a header

Required Methods

fn parent_hash(&self) -> B256

Retrieves the parent hash of the block

fn ommers_hash(&self) -> B256

Retrieves the ommers hash of the block

fn beneficiary(&self) -> Address

Retrieves the beneficiary (miner) of the block

fn state_root(&self) -> B256

Retrieves the state root hash of the block

fn transactions_root(&self) -> B256

Retrieves the transactions root hash of the block

fn receipts_root(&self) -> B256

Retrieves the receipts root hash of the block

fn withdrawals_root(&self) -> Option<B256>

Retrieves the withdrawals root hash of the block, if available

fn logs_bloom(&self) -> Bloom

Retrieves the logs bloom filter of the block

fn difficulty(&self) -> U256

Retrieves the difficulty of the block

fn number(&self) -> BlockNumber

Retrieves the block number

fn gas_limit(&self) -> u64

Retrieves the gas limit of the block

fn gas_used(&self) -> u64

Retrieves the gas used by the block

fn timestamp(&self) -> u64

Retrieves the timestamp of the block

fn mix_hash(&self) -> B256

Retrieves the mix hash of the block

fn nonce(&self) -> B64

Retrieves the nonce of the block

fn base_fee_per_gas(&self) -> Option<u64>

Retrieves the base fee per gas of the block, if available

fn blob_gas_used(&self) -> Option<u64>

Retrieves the blob gas used by the block, if available

fn excess_blob_gas(&self) -> Option<u64>

Retrieves the excess blob gas of the block, if available

fn parent_beacon_block_root(&self) -> Option<B256>

Retrieves the parent beacon block root of the block, if available

fn requests_root(&self) -> Option<B256>

Retrieves the requests root of the block, if available

fn extra_data(&self) -> &Bytes

Retrieves the block’s extra data field