alloy_consensus

Trait RlpEncodableReceipt

pub trait RlpEncodableReceipt {
    // Required methods
    fn rlp_encoded_length_with_bloom(&self, bloom: &Bloom) -> usize;
    fn rlp_encode_with_bloom(&self, bloom: &Bloom, out: &mut dyn BufMut);
}

Receipt type that knows how to encode itself with a Bloom value.

Required Methods

fn rlp_encoded_length_with_bloom(&self, bloom: &Bloom) -> usize

Returns the length of the receipt payload with the provided bloom filter.

fn rlp_encode_with_bloom(&self, bloom: &Bloom, out: &mut dyn BufMut)

RLP encodes the receipt with the provided bloom filter.