alloy_consensus

Trait TxReceipt

pub trait TxReceipt:
    Clone
    + Debug
    + PartialEq
    + Eq
    + Send
    + Sync {
    type Log;

    // Required methods
    fn status_or_post_state(&self) -> Eip658Value;
    fn status(&self) -> bool;
    fn bloom(&self) -> Bloom;
    fn cumulative_gas_used(&self) -> u64;
    fn logs(&self) -> &[Self::Log];

    // Provided methods
    fn bloom_cheap(&self) -> Option<Bloom> { ... }
    fn with_bloom_ref(&self) -> ReceiptWithBloom<&Self> { ... }
    fn into_with_bloom(self) -> ReceiptWithBloom<Self> { ... }
    fn into_with_bloom_unchecked(
        self,
        logs_bloom: Bloom,
    ) -> ReceiptWithBloom<Self> { ... }
    fn into_logs(self) -> Vec<Self::Log>
       where Self::Log: Clone { ... }
}

Receipt is the result of a transaction execution.

Required Associated Types

type Log

The associated log type.

Required Methods

fn status_or_post_state(&self) -> Eip658Value

Returns the status or post state of the transaction.

Note

Use this method instead of TxReceipt::status when the transaction is pre-EIP-658.

fn status(&self) -> bool

Returns true if the transaction was successful OR if the transaction is pre-EIP-658. Results for transactions before EIP-658 are not reliable.

Note

Caution must be taken when using this method for deep-historical receipts, as it may not accurately reflect the status of the transaction. The transaction status is not knowable from the receipt for transactions before EIP-658.

This can be handled using TxReceipt::status_or_post_state.

fn bloom(&self) -> Bloom

Returns the bloom filter for the logs in the receipt. This operation may be expensive.

fn cumulative_gas_used(&self) -> u64

Returns the cumulative gas used in the block after this transaction was executed.

fn logs(&self) -> &[Self::Log]

Returns the logs emitted by this transaction.

Provided Methods

fn bloom_cheap(&self) -> Option<Bloom>

Returns the bloom filter for the logs in the receipt, if it is cheap to compute.

fn with_bloom_ref(&self) -> ReceiptWithBloom<&Self>

Returns ReceiptWithBloom with the computed bloom filter Self::bloom and a reference to the receipt.

fn into_with_bloom(self) -> ReceiptWithBloom<Self>

Consumes the type and converts it into ReceiptWithBloom with the computed bloom filter Self::bloom and the receipt.

fn into_with_bloom_unchecked(self, logs_bloom: Bloom) -> ReceiptWithBloom<Self>

Consumes the type and converts it into ReceiptWithBloom with the given bloom filter.

fn into_logs(self) -> Vec<Self::Log>
where Self::Log: Clone,

Consumes the type and returns the logs emitted by this transaction as a vector.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.