alloy_consensus

Module transaction

Transaction types.

Re-exports

pub use eip4844::TxEip4844;

pub use eip4844::TxEip4844Variant;

pub use eip4844::TxEip4844WithSidecar;

pub use pooled::PooledTransaction;

Modules

eip4844

[EIP-4844] constants, helpers, and types.

eip4844_utils

Utilities for working with EIP-4844 field elements and implementing SidecarCoder.

pooled

Defines the exact transaction variant that are allowed to be propagated over the eth p2p protocol.

signed_legacy_serde

Helper module for encoding signatures of transactions wrapped into Signed in legacy format.

untagged_legacy_serde

Helper module for deserializing legacy transactions and ensuring that the type field is not present.

Structs

BlobTransactionSidecar

This represents a set of blobs, and its corresponding commitments and proofs.

Recovered

Signed object with recovered signer.

SidecarBuilder

Build a BlobTransactionSidecar from an arbitrary amount of data.

SimpleCoder

Simple coder that only uses the last 31 bytes of each blob. This is the default coder for the SidecarBuilder.

TransactionInfo

Additional fields in the context of a (maybe) pending block that contains this transaction.

TransactionMeta

Additional fields in the context of a block that contains this mined transaction.

TxEip1559

A transaction with a priority fee (EIP-1559).

TxEip2930

Transaction with an AccessList (EIP-2930).

TxEip7702

A transaction with a priority fee (EIP-7702).

TxLegacy

Legacy transaction.

Enums

Either

Re-export for convenience The enum Either with variants Left and Right is a general purpose sum type with two cases.

EthereumTxEnvelope

The Ethereum EIP-2718 Transaction Envelope.

EthereumTypedTransaction

The TypedTransaction enum represents all Ethereum transaction request types.

TxType

Transaction types supported by EthereumTxEnvelope.

Traits

SidecarCoder

A strategy for coding and decoding data into sidecars.

SignableTransaction

A signable transaction.

SignerRecoverable

A type that can recover the signer of a transaction.

Transaction

Represents a minimal EVM transaction. Currently, EIP-1559, EIP-4844, and EIP-7702 support dynamic fees. We call these transactions “dynamic fee transactions”. We call non dynamic fee transactions(EIP-155, EIP-2930) “legacy fee transactions”.

TransactionEnvelope

A typed transaction envelope.

TxEip4844Sidecar

An EIP-4844 transaction sidecar.

Functions

from_eip155_value

Helper for decoding EIP-155 v value into y_parity boolean and optional chain_id.

to_eip155_value

Helper for encoding y_parity boolean and optional chain_id into EIP-155 v value.

Type Aliases

Blob

A Blob serialized as 0x-prefixed hex string

Bytes48

A commitment/proof serialized as 0x-prefixed hex string

TxEnvelope

The Ethereum EIP-2718 Transaction Envelope.

TypedTransaction

Basic typed transaction which can contain both TxEip4844 and TxEip4844WithSidecar.