Module alloy_consensus::transaction::eip4844

Re-exports

pub use builder::SidecarBuilder;

pub use builder::SidecarCoder;

pub use builder::SimpleCoder;

Modules

builder 🔒

utils

Utilities for working with EIP-4844 field elements and implementing SidecarCoder.

Structs

BlobTransactionSidecar

This represents a set of blobs, and its corresponding commitments and proofs.

TxEip4844

EIP-4844 Blob Transaction

TxEip4844WithSidecar

EIP-4844 Blob Transaction

Enums

BlobTransactionValidationErrorkzg

An error that can occur when validating a BlobTransactionSidecar::validate.

TxEip4844Variant

EIP-4844 Blob Transaction

Type Aliases

Blob

A Blob serialized as 0x-prefixed hex string

Bytes48

A commitment/proof serialized as 0x-prefixed hex string