alloy_consensus::transaction

Enum EthereumTxEnvelope

pub enum EthereumTxEnvelope<Eip4844> {
    Legacy(Signed<TxLegacy>),
    Eip2930(Signed<TxEip2930>),
    Eip1559(Signed<TxEip1559>),
    Eip4844(Signed<Eip4844>),
    Eip7702(Signed<TxEip7702>),
}

The Ethereum EIP-2718 Transaction Envelope.

Note:

This enum distinguishes between tagged and untagged legacy transactions, as the in-protocol merkle tree may commit to EITHER 0-prefixed or raw. Therefore we must ensure that encoding returns the precise byte-array that was decoded, preserving the presence or absence of the TransactionType flag.

Variants

Legacy(Signed<TxLegacy>)

An untagged TxLegacy.

Eip2930(Signed<TxEip2930>)

A TxEip2930 tagged with type 1.

Eip1559(Signed<TxEip1559>)

A TxEip1559 tagged with type 2.

Eip4844(Signed<Eip4844>)

A TxEip4844 tagged with type 3. An EIP-4844 transaction has two network representations: 1 - The transaction itself, which is a regular RLP-encoded transaction and used to retrieve historical transactions..

2 - The transaction with a sidecar, which is the form used to send transactions to the network.

Eip7702(Signed<TxEip7702>)

A TxEip7702 tagged with type 4.

Implementations

impl<T: Encodable7594> EthereumTxEnvelope<TxEip4844Variant<T>>

pub fn try_into_pooled( self, ) -> Result<EthereumTxEnvelope<TxEip4844WithSidecar<T>>, ValueError<Self>>

Attempts to convert the envelope into the pooled variant.

Returns an error if the envelope’s variant is incompatible with the pooled format: crate::TxEip4844 without the sidecar.

impl EthereumTxEnvelope<TxEip4844>

pub fn try_into_pooled<T>( self, ) -> Result<EthereumTxEnvelope<TxEip4844WithSidecar<T>>, ValueError<Self>>

Attempts to convert the envelope into the pooled variant.

Returns an error if the envelope’s variant is incompatible with the pooled format: crate::TxEip4844 without the sidecar.

pub fn try_into_pooled_eip4844<T>( self, sidecar: T, ) -> Result<EthereumTxEnvelope<TxEip4844WithSidecar<T>>, ValueError<Self>>

Converts from an EIP-4844 transaction to a EthereumTxEnvelope<TxEip4844WithSidecar<T>> with the given sidecar.

Returns an Err containing the original EthereumTxEnvelope if the transaction is not an EIP-4844 variant.

impl<T> EthereumTxEnvelope<T>

pub fn new_unchecked( transaction: EthereumTypedTransaction<T>, signature: Signature, hash: B256, ) -> Self
where T: RlpEcdsaEncodableTx,

Creates a new signed transaction from the given transaction, signature and hash.

Caution: This assumes the given hash is the correct transaction hash.

pub fn new( transaction: EthereumTypedTransaction<T>, signature: Signature, hash: B256, ) -> Self
where T: RlpEcdsaEncodableTx,

👎Deprecated: Use new_unchecked() instead

Creates a new signed transaction from the given transaction, signature and hash.

Caution: This assumes the given hash is the correct transaction hash.

pub fn new_unhashed( transaction: EthereumTypedTransaction<T>, signature: Signature, ) -> Self
where T: RlpEcdsaEncodableTx + SignableTransaction<Signature>,

Creates a new signed transaction from the given typed transaction and signature without the hash.

Note: this only calculates the hash on the first EthereumTxEnvelope::hash call.

pub fn into_typed_transaction(self) -> EthereumTypedTransaction<T>
where T: RlpEcdsaEncodableTx,

Consumes the type, removes the signature and returns the transaction.

impl<Eip4844> EthereumTxEnvelope<Eip4844>

pub fn map_eip4844<U>( self, f: impl FnMut(Eip4844) -> U, ) -> EthereumTxEnvelope<U>

Converts the EIP-4844 variant of this transaction with the given closure.

This is intended to convert between the EIP-4844 variants, specifically for stripping away non consensus data (blob sidecar data).

pub const fn tx_type(&self) -> TxType

Return the TxType of the inner txn.

pub fn into_signed(self) -> Signed<EthereumTypedTransaction<Eip4844>>
where EthereumTypedTransaction<Eip4844>: From<Eip4844>,

Consumes the type into a Signed

impl<Eip4844: RlpEcdsaEncodableTx> EthereumTxEnvelope<Eip4844>

pub const fn is_legacy(&self) -> bool

Returns true if the transaction is a legacy transaction.

pub const fn is_eip2930(&self) -> bool

Returns true if the transaction is an EIP-2930 transaction.

pub const fn is_eip1559(&self) -> bool

Returns true if the transaction is an EIP-1559 transaction.

pub const fn is_eip4844(&self) -> bool

Returns true if the transaction is an EIP-4844 transaction.

pub const fn is_eip7702(&self) -> bool

Returns true if the transaction is an EIP-7702 transaction.

pub const fn is_replay_protected(&self) -> bool

Returns true if the transaction is replay protected.

All non-legacy transactions are replay protected, as the chain id is included in the transaction body. Legacy transactions are considered replay protected if the v value is not 27 or 28, according to the rules of EIP-155.

pub const fn as_legacy(&self) -> Option<&Signed<TxLegacy>>

Returns the TxLegacy variant if the transaction is a legacy transaction.

pub const fn as_eip2930(&self) -> Option<&Signed<TxEip2930>>

Returns the TxEip2930 variant if the transaction is an EIP-2930 transaction.

pub const fn as_eip1559(&self) -> Option<&Signed<TxEip1559>>

Returns the TxEip1559 variant if the transaction is an EIP-1559 transaction.

pub const fn as_eip4844(&self) -> Option<&Signed<Eip4844>>

Returns the TxEip4844Variant variant if the transaction is an EIP-4844 transaction.

pub const fn as_eip7702(&self) -> Option<&Signed<TxEip7702>>

Returns the TxEip7702 variant if the transaction is an EIP-7702 transaction.

pub fn signature_hash(&self) -> B256
where Eip4844: SignableTransaction<Signature>,

Calculate the signing hash for the transaction.

pub const fn signature(&self) -> &Signature

Return the reference to signature.

pub fn tx_hash(&self) -> &B256

Return the hash of the inner Signed.

pub fn hash(&self) -> &B256

Reference to transaction hash. Used to identify transaction.

pub fn eip2718_encoded_length(&self) -> usize

Return the length of the inner txn, including type byte length

impl<T: Encodable7594> EthereumTxEnvelope<TxEip4844WithSidecar<T>>

pub fn into_envelope(self) -> EthereumTxEnvelope<TxEip4844Variant<T>>

Converts the transaction into EthereumTxEnvelope<TxEip4844Variant<T>>.