Enum alloy_consensus::transaction::TxEnvelope

pub enum TxEnvelope {
    Legacy(Signed<TxLegacy>),
    Eip2930(Signed<TxEip2930>),
    Eip1559(Signed<TxEip1559>),
    Eip4844(Signed<TxEip4844Variant>),
    Eip7702(Signed<TxEip7702>),
}

The Ethereum EIP-2718 Transaction Envelope.

Note:

This enum distinguishes between tagged and untagged legacy transactions, as the in-protocol merkle tree may commit to EITHER 0-prefixed or raw. Therefore we must ensure that encoding returns the precise byte-array that was decoded, preserving the presence or absence of the TransactionType flag.

Variants

Legacy(Signed<TxLegacy>)

An untagged TxLegacy.

Eip2930(Signed<TxEip2930>)

A TxEip2930 tagged with type 1.

Eip1559(Signed<TxEip1559>)

A TxEip1559 tagged with type 2.

Eip4844(Signed<TxEip4844Variant>)

A TxEip4844 tagged with type 3. An EIP-4844 transaction has two network representations: 1 - The transaction itself, which is a regular RLP-encoded transaction and used to retrieve historical transactions..

2 - The transaction with a sidecar, which is the form used to send transactions to the network.

Eip7702(Signed<TxEip7702>)

A TxEip7702 tagged with type 4.

Implementations

impl TxEnvelope

pub const fn is_legacy(&self) -> bool

Returns true if the transaction is a legacy transaction.

pub const fn is_eip2930(&self) -> bool

Returns true if the transaction is an EIP-2930 transaction.

pub const fn is_eip1559(&self) -> bool

Returns true if the transaction is an EIP-1559 transaction.

pub const fn is_eip4844(&self) -> bool

Returns true if the transaction is an EIP-4844 transaction.

pub const fn is_eip7702(&self) -> bool

Returns true if the transaction is an EIP-7702 transaction.

pub const fn is_replay_protected(&self) -> bool

Returns true if the transaction is replay protected.

All non-legacy transactions are replay protected, as the chain id is included in the transaction body. Legacy transactions are considered replay protected if the v value is not 27 or 28, according to the rules of EIP-155.

pub const fn as_legacy(&self) -> Option<&Signed<TxLegacy>>

Returns the TxLegacy variant if the transaction is a legacy transaction.

pub const fn as_eip2930(&self) -> Option<&Signed<TxEip2930>>

Returns the TxEip2930 variant if the transaction is an EIP-2930 transaction.

pub const fn as_eip1559(&self) -> Option<&Signed<TxEip1559>>

Returns the TxEip1559 variant if the transaction is an EIP-1559 transaction.

pub const fn as_eip4844(&self) -> Option<&Signed<TxEip4844Variant>>

Returns the TxEip4844Variant variant if the transaction is an EIP-4844 transaction.

pub const fn as_eip7702(&self) -> Option<&Signed<TxEip7702>>

Returns the TxEip7702 variant if the transaction is an EIP-7702 transaction.

pub fn signature_hash(&self) -> B256

Calculate the signing hash for the transaction.

pub const fn signature(&self) -> &Signature

Return the reference to signature.

pub const fn tx_hash(&self) -> &B256

Return the hash of the inner Signed.

pub const fn tx_type(&self) -> TxType

Return the TxType of the inner txn.

pub fn eip2718_encoded_length(&self) -> usize

Return the length of the inner txn, including type byte length