Struct alloy_consensus::transaction::legacy::TxLegacy

pub struct TxLegacy {
    pub chain_id: Option<ChainId>,
    pub nonce: u64,
    pub gas_price: u128,
    pub gas_limit: u64,
    pub to: TxKind,
    pub value: U256,
    pub input: Bytes,
}

Legacy transaction.

Fields
chain_id: Option<ChainId>

Added as EIP-155: Simple replay attack protection

nonce: u64

A scalar value equal to the number of transactions sent by the sender; formally Tn.

gas_price: u128

A scalar value equal to the number of Wei to be paid per unit of gas for all computation costs incurred as a result of the execution of this transaction; formally Tp.

As ethereum circulation is around 120mil eth as of 2022 that is around 120000000000000000000000000 wei we are safe to use u128 as its max number is: 340282366920938463463374607431768211455

gas_limit: u64

A scalar value equal to the maximum amount of gas that should be used in executing this transaction. This is paid up-front, before any computation is done and may not be increased later; formally Tg.

to: TxKind

The 160-bit address of the message call’s recipient or, for a contract creation transaction, ∅, used here to denote the only member of B0 ; formally Tt.

value: U256

A scalar value equal to the number of Wei to be transferred to the message call’s recipient or, in the case of contract creation, as an endowment to the newly created account; formally Tv.

input: Bytes

Input has two uses depending if transaction is Create or Call (if to field is None or Some). pub init: An unlimited size byte array specifying the EVM-code for the account initialisation procedure CREATE, data: An unlimited size byte array specifying the input data of the message call, formally Td.

Implementations

impl TxLegacy

pub const TX_TYPE: isize = 0isize

The EIP-2718 transaction type.

pub fn size(&self) -> usize

Calculates a heuristic for the in-memory size of the TxLegacy transaction.

pub(crate) fn fields_len(&self) -> usize

Outputs the length of the transaction’s fields, without a RLP header or length of the eip155 fields.

pub(crate) fn encode_fields(&self, out: &mut dyn BufMut)

Encodes only the transaction’s fields into the desired buffer, without a RLP header or eip155 fields.

pub fn encode_with_signature_fields( &self, signature: &Signature, out: &mut dyn BufMut )

Encodes the transaction from RLP bytes, including the signature. This does not encode a tx type byte or string header.

This does encode a list header and include a signature.

pub(crate) fn encoded_len_with_signature(&self, signature: &Signature) -> usize

Returns what the encoded length should be, if the transaction were RLP encoded with the given signature.

pub(crate) fn encode_eip155_signing_fields(&self, out: &mut dyn BufMut)

Encodes EIP-155 arguments into the desired buffer. Only encodes values for legacy transactions.

pub(crate) fn eip155_fields_len(&self) -> usize

Outputs the length of EIP-155 fields. Only outputs a non-zero value for EIP-155 legacy transactions.

pub(crate) fn decode_signed_fields(buf: &mut &[u8]) -> Result<Signed<Self>>

Decodes the transaction from RLP bytes, including the signature.

This does not expect the bytes to start with a transaction type byte or string header.

This does expect the bytes to start with a list header and include a signature.

pub(crate) fn decode_fields(data: &mut &[u8]) -> Result<Self>

Decode the RLP fields of the transaction, without decoding an RLP header.