alloy_consensus::transaction

Struct Recovered

pub struct Recovered<T> { /* private fields */ }

Signed object with recovered signer.

Implementations

impl<T> Recovered<T>

pub const fn signer(&self) -> Address

Signer of the object recovered from signature

pub const fn signer_ref(&self) -> &Address

Reference to the signer of the object recovered from signature

pub const fn inner(&self) -> &T

Reference to the inner recovered object.

pub const fn inner_mut(&mut self) -> &mut T

Reference to the inner recovered object.

pub fn into_inner(self) -> T

Reference to the inner signed object.

pub fn clone_inner(&self) -> T
where T: Clone,

Clone the inner signed object.

pub const fn tx(&self) -> &T

๐Deprecated: Use inner instead

Returns a reference to the transaction.

pub fn into_tx(self) -> T

๐Deprecated: Use into_inner instead

Transform back to the transaction.

pub fn clone_tx(&self) -> T
where T: Clone,

๐Deprecated: Use clone_inner instead

Clone the inner transaction.

pub fn into_parts(self) -> (T, Address)

Dissolve Self to its component

pub const fn as_recovered_ref(&self) -> Recovered<&T>

Converts from &Recovered<T> to Recovered<&T>.

pub const fn new_unchecked(inner: T, signer: Address) -> Self

Create Recovered from the given transaction and Address of the signer.

Note: This does not check if the signer is the actual signer of the transaction.

pub fn convert<Tx>(self) -> Recovered<Tx>
where Tx: From<T>,

Converts the inner signed object to the given alternative that is From<T>

pub fn convert_transaction<Tx>(self) -> Recovered<Tx>
where Tx: From<T>,

๐Deprecated: Use convert_inner instead

Converts the transaction type to the given alternative that is From<T>

pub fn try_convert<Tx, E>(self) -> Result<Recovered<Tx>, Tx::Error>
where Tx: TryFrom<T>,

Converts the inner signed object to the given alternative that is TryFrom<T>

pub fn try_convert_transaction<Tx, E>(self) -> Result<Recovered<Tx>, Tx::Error>
where Tx: TryFrom<T>,

๐Deprecated: Use try_convert_inner instead

Converts the transaction to the given alternative that is TryFrom<T>

pub fn map<Tx>(self, f: impl FnOnce(T) -> Tx) -> Recovered<Tx>

Applies the given closure to the inner signed object.

pub fn map_transaction<Tx>(self, f: impl FnOnce(T) -> Tx) -> Recovered<Tx>

๐Deprecated: Use map_inner instead

Applies the given closure to the inner transaction type.

pub fn try_map<Tx, E>( self, f: impl FnOnce(T) -> Result<Tx, E>, ) -> Result<Recovered<Tx>, E>

Applies the given fallible closure to the inner signed object.

pub fn try_map_transaction<Tx, E>( self, f: impl FnOnce(T) -> Result<Tx, E>, ) -> Result<Recovered<Tx>, E>

๐Deprecated: Use try_map_inner instead

Applies the given fallible closure to the inner transaction type.

pub fn into_encoded_with(self, encoding: impl Into<Bytes>) -> WithEncoded<Self>

Returns the WithEncoded representation of Recovered with the given encoding.

pub fn into_encoded(self) -> WithEncoded<Self>
where T: Encodable2718,

Encodes the inner type and returns the WithEncoded representation of Recovered.

impl<T> Recovered<&T>

pub fn cloned(self) -> Recovered<T>
where T: Clone,

Maps a Recovered<&T> to a Recovered<T> by cloning the transaction.

pub const fn copied(&self) -> Self

Helper function to explicitly create a new copy of Recovered<&T>