alloy_consensus::transaction

Trait TxHashable

pub trait TxHashable<S>: Typed2718 {
    // Required method
    fn tx_hash_with_type(&self, signature: &S, ty: u8) -> TxHash;

    // Provided method
    fn tx_hash(&self, signature: &S) -> TxHash { ... }
}

Generic trait to get a transaction hash from any signature type

Required Methods

fn tx_hash_with_type(&self, signature: &S, ty: u8) -> TxHash

Calculate the transaction hash for the given signature and type.

Provided Methods

fn tx_hash(&self, signature: &S) -> TxHash

Calculate the transaction hash for the given signature.