Module alloy_consensus::transaction::eip4844

Re-exports

pub use builder::SidecarBuilder;

pub use builder::SidecarCoder;

pub use builder::SimpleCoder;

Modules

builder 🔒

utils

Utilities for working with EIP-4844 field elements and implementing SidecarCoder.

Structs

Blobkzg

A basic blob data.

BlobTransactionSidecar

This represents a set of blobs, and its corresponding commitments and proofs.

BlobTransactionSidecarRlp 🔒

Bytes48kzg

An array of 48 bytes. Represents an untrusted (potentially invalid) commitment/proof.

TxEip4844

EIP-4844 Blob Transaction

TxEip4844WithSidecar

EIP-4844 Blob Transaction

Enums

BlobTransactionValidationErrorkzg

An error that can occur when validating a TxEip4844Variant.

TxEip4844Variant

EIP-4844 Blob Transaction

Functions

kzg_to_versioned_hash 🔒

Calculates the versioned hash for a KzgCommitment